GUI components look and feel improved

+++ src/com/alife/screenshot/DirectoryHelper.java
@@ -6,23 +6,24 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Scanner;
 
+/**
+ * A class to manage directory for screenshot selection.
+ *
+ * @author Santosh Sagar
+ */
+
 public class DirectoryHelper {
 
 	private static String dirPath;
-	private static final String dirPathFileName=".scrDirectory"; 
-	
+	private static final String dirPathFileName=".screenShotTaker";
+
 	public DirectoryHelper() {
 		dirPath="";
 		dirPath=System.getProperty("user.home");
 		if(!dirPath.equals(""))
 			dirPath=dirPath+"\\";
 	}
-	
-	/*public static void main(String[] args) throws IOException {
-		DirectoryHelper dh=new DirectoryHelper();
-		dh.setDir("d:/");
-		System.out.println(dh.getDir());
-	}*/
+
 	public String getDir() throws FileNotFoundException{
 		String result="*";
 		File file=new File(dirPath+dirPathFileName);
@@ -31,24 +32,26 @@ public class DirectoryHelper {
 			result=in.nextLine();
 		in.close();
 		return result;
-			
+
 	}
-	
-	/*
-	 * This Method will store the selected directory in user/Home/.directory folder. 
-	 * 
-	 * @Param dir String
-	 * @author Santosh Sagar
-	 * @version 1.0.0
+
+	/**
+	 * To store the selected directory in user/home/.screenShotTaker folder.
+	 *
 	 */
 	public int setDir(String dir) throws IOException{
 		int result=0;
+		dirPath="C:\\Program Files (x86)\\Adobe";
 		File file=new File(dirPath+dirPathFileName);
+		if( !file.exists() && !file.canWrite() ) {
+			System.out.println("File doesn't exist and write permission is not allowd");
+			throw new IOException("Permission denied for path : "+file.getAbsolutePath());
+		}
 		FileWriter out=new FileWriter(file);
 		out.write(dir);
 		result=1;
 		out.close();
 		return result;
-			
+
 	}
 }
